+++ src/main/java/com/box/sdk/BoxFile.java
@@ -846,6 +846,20 @@ public class BoxFile extends BoxItem {
     }
 
     /**
+     * Uploads a new version of this file, replacing the current version. Note that only users with premium accounts
+     * will be able to view and recover previous versions of the file.
+     *
+     * @param fileContent     a stream containing the new file contents.
+     * @param fileContentSHA1 a string containing the SHA1 hash of the new file contents.
+     * @param modified        the date that the new version was modified.
+     * @param name            the new name for the file
+     * @return the uploaded file version.
+     */
+    public BoxFile.Info uploadNewVersion(InputStream fileContent, String fileContentSHA1, Date modified, String name) {
+        return this.uploadNewVersion(fileContent, fileContentSHA1, modified, name, 0, null);
+    }
+
+    /**
      * Uploads a new version of this file, replacing the current version, while reporting the progress to a
      * ProgressListener. Note that only users with premium accounts will be able to view and recover previous versions
      * of the file.
@@ -875,6 +889,24 @@ public class BoxFile extends BoxItem {
      */
     public BoxFile.Info uploadNewVersion(InputStream fileContent, String fileContentSHA1, Date modified, long fileSize,
                               ProgressListener listener) {
+        return this.uploadNewVersion(fileContent, fileContentSHA1, modified, null, fileSize, listener);
+    }
+
+    /**
+     * Uploads a new version of this file, replacing the current version, while reporting the progress to a
+     * ProgressListener. Note that only users with premium accounts will be able to view and recover previous versions
+     * of the file.
+     *
+     * @param fileContent     a stream containing the new file contents.
+     * @param fileContentSHA1 the SHA1 hash of the file contents. will be sent along in the Content-MD5 header
+     * @param modified        the date that the new version was modified.
+     * @param name            the new name for the file
+     * @param fileSize        the size of the file used for determining the progress of the upload.
+     * @param listener        a listener for monitoring the upload's progress.
+     * @return the uploaded file version.
+     */
+    public BoxFile.Info uploadNewVersion(InputStream fileContent, String fileContentSHA1, Date modified, String name,
+                                         long fileSize, ProgressListener listener) {
         URL uploadURL = CONTENT_URL_TEMPLATE.build(this.getAPI().getBaseUploadURL(), this.getID());
         BoxMultipartRequest request = new BoxMultipartRequest(getAPI(), uploadURL);
 
@@ -888,9 +920,16 @@ public class BoxFile extends BoxItem {
             request.setContentSHA1(fileContentSHA1);
         }
 
+        JsonObject attributesJSON = new JsonObject();
         if (modified != null) {
-            request.putField("content_modified_at", modified);
+            attributesJSON.add("content_modified_at", BoxDateFormat.format(modified));
         }
+
+        if (name != null) {
+            attributesJSON.add("name", name);
+        }
+
+        request.putField("attributes", attributesJSON.toString());
 
         BoxJSONResponse response;
         if (listener == null) {
@@ -1586,6 +1625,7 @@ public class BoxFile extends BoxItem {
         private List<String> allowedInviteeRoles;
         private Boolean hasCollaborations;
         private String uploaderDisplayName;
+        private BoxClassification classification;
 
         /**
          * Constructs an empty Info object.
@@ -1766,6 +1806,14 @@ public class BoxFile extends BoxItem {
             return this.uploaderDisplayName;
         }
 
+        /**
+         * Gets the metadata classification type of this file.
+         * @return the metadata classification type of this file.
+         */
+        public BoxClassification getClassification() {
+            return this.classification;
+        }
+
         @Override
         protected void parseJSONMember(JsonObject.Member member) {
             super.parseJSONMember(member);
@@ -1817,6 +1865,12 @@ public class BoxFile extends BoxItem {
                     this.representations = Parsers.parseRepresentations(jsonObject);
                 } else if (memberName.equals("uploader_display_name")) {
                     this.uploaderDisplayName = value.asString();
+                } else if (memberName.equals("classification")) {
+                    if (value.isNull()) {
+                        this.classification = null;
+                    } else {
+                        this.classification = new BoxClassification(value.asObject());
+                    }
                 }
             } catch (Exception e) {
                 throw new BoxDeserializationException(memberName, value.toString(), e);